Struct switchboard_solana::attestation_program::accounts::verifier::VerifierAccountData

#[repr(packed(1))]
pub struct VerifierAccountData {
    pub enclave: Quote,
    pub authority: Pubkey,
    pub attestation_queue: Pubkey,
    pub created_at: i64,
    pub is_on_queue: bool,
    pub last_heartbeat: i64,
    pub reward_escrow: Pubkey,
    pub stake_wallet: Pubkey,
    pub success_ewma: Ewma,
    pub failure_ewma: Ewma,
    pub gateway_uri: [u8; 64],
    pub _ebuf: [u8; 872],
}

Fields
enclave: Quote

Represents the state of the quote verifiers enclave.

authority: Pubkey

The authority of the EnclaveAccount which is permitted to make account changes.

attestation_queue: Pubkey

Queue used for attestation to verify a MRENCLAVE measurement.

created_at: i64

The unix timestamp when the quote was created.

is_on_queue: bool

Whether the quote is located on the AttestationQueues buffer.

last_heartbeat: i64

The last time the quote heartbeated on-chain.

reward_escrow: Pubkey

The SwitchboardWallet account containing the reward escrow for verifying quotes on-chain. We should set this whenever the operator changes so we dont need to pass another account and can verify with has_one.

stake_wallet: Pubkey

The SwitchboardWallet account containing the queues required min_stake. Needs to be separate from the reward_escrow. Allows easier 3rd party management of stake from rewards.

success_ewma: Ewma

Tracks the number of successful price attestations.

failure_ewma: Ewma

Tracks the number of failed price attestations (user reported).

gateway_uri: [u8; 64]

URI location of the verifier’s gateway.

_ebuf: [u8; 872]

Reserved.

Implementations

impl VerifierAccountData

pub fn size() -> usize

pub fn new<'info>( quote_account_info: &'info AccountInfo<'info> ) -> Result<Ref<'info, VerifierAccountData>>

Returns the deserialized Switchboard Quote account

Arguments

quote_account_info - A Solana AccountInfo referencing an existing Switchboard QuoteAccount

Examples

use switchboard_solana::VerifierAccountData;

let quote_account = VerifierAccountData::new(quote_account_info)?;

pub fn new_from_bytes(data: &[u8]) -> Result<&VerifierAccountData>

Returns the deserialized Switchboard Quote account

Arguments

data - A Solana AccountInfo’s data buffer

Examples

use switchboard_solana::VerifierAccountData;

let quote_account = VerifierAccountData::new(quote_account_info.try_borrow_data()?)?;

pub fn signer(&self) -> Pubkey

pub fn assert_signer(&self, signer: &AccountInfo<'_>) -> Result<()>

pub fn is_verified(&self, clock: &Clock) -> bool

pub fn verify(&self, clock: &Clock) -> Result<()>

pub fn gateway_uri(&self) -> Option<String>

pub fn ed25519_signer(&self) -> Option<Pubkey>

pub fn secp256k1_signer(&self) -> Option<[u8; 64]>

pub fn libsecp256k1_signer(&self) -> Option<PublicKey>

pub fn fetch(client: &RpcClient, pubkey: Pubkey) -> Result<Self, SbError>

pub async fn fetch_async( client: &RpcClient, pubkey: Pubkey ) -> Result<Self, SbError>

pub fn fetch_sync<T: SyncClient>( client: &T, pubkey: Pubkey ) -> Result<Self, SbError>