Struct switchboard_solana::attestation_program::instructions::request_init_and_trigger::FunctionRequestInitAndTrigger

pub struct FunctionRequestInitAndTrigger<'info> {
    pub request: AccountInfo<'info>,
    pub authority: AccountInfo<'info>,
    pub function: AccountInfo<'info>,
    pub function_authority: Option<AccountInfo<'info>>,
    pub escrow: AccountInfo<'info>,
    pub mint: AccountInfo<'info>,
    pub state: AccountInfo<'info>,
    pub attestation_queue: AccountInfo<'info>,
    pub payer: AccountInfo<'info>,
    pub system_program: AccountInfo<'info>,
    pub token_program: AccountInfo<'info>,
    pub associated_token_program: AccountInfo<'info>,
}

Fields
request: AccountInfo<'info>authority: AccountInfo<'info>function: AccountInfo<'info>function_authority: Option<AccountInfo<'info>>

CHECK: function authority required to permit new requests

escrow: AccountInfo<'info>mint: AccountInfo<'info>state: AccountInfo<'info>attestation_queue: AccountInfo<'info>payer: AccountInfo<'info>system_program: AccountInfo<'info>token_program: AccountInfo<'info>associated_token_program: AccountInfo<'info>
Implementations

impl<'info> FunctionRequestInitAndTrigger<'info>

pub fn get_instruction( &self, program_id: Pubkey, params: FunctionRequestInitAndTriggerParams ) -> Result<Instruction>

pub fn invoke( &self, program: AccountInfo<'info>, bounty: Option<u64>, slots_until_expiration: Option<u64>, max_container_params_len: Option<u32>, container_params: Option<Vec<u8>>, garbage_collection_slot: Option<u64>, valid_after_slot: Option<u64> ) -> ProgramResult

Invokes the instruction.

Arguments

program: The Switchboard Attestation program AccountInfo.

max_container_params_len: The maximum length of the vector containing the container parameters.

Default: 256 bytes.

Example: Some(512).

container_params: The parameters for the container.

Default: Empty vector.

Example: Some(request_params.into_bytes()).

garbage_collection_slot: The slot when the request can be closed by anyone and is considered expired.

Default: None, meaning only the authority can close the request.

bounty: An optional fee to reward oracles for priority processing.

Default: 0 lamports.

slots_until_expiration: An optional specification for the max number of slots the request can be processed in.

Default: 2250 slots (roughly 15 minutes at 400 ms/slot).

Minimum: 150 slots (approximately 1 minute at 400 ms/slot).

valid_after_slot: Schedule when the request should be valid for processing.

Default: 0 slots, meaning the request is valid immediately for oracles to process.

Returns

ProgramResult: Indicates the result of the instruction execution.

pub fn invoke_signed( &self, program: AccountInfo<'info>, bounty: Option<u64>, slots_until_expiration: Option<u64>, max_container_params_len: Option<u32>, container_params: Option<Vec<u8>>, garbage_collection_slot: Option<u64>, valid_after_slot: Option<u64>, signer_seeds: &[&[&[u8]]] ) -> ProgramResult

Invokes the instruction using a signed authority.

Arguments

program: The Switchboard Attestation program AccountInfo.

max_container_params_len: The maximum length of the vector containing the container parameters.

Default: 256 bytes.

Example: Some(512).

container_params: The parameters for the container.

Default: Empty vector.

Example: Some(request_params.into_bytes()).

garbage_collection_slot: The slot when the request can be closed by anyone and is considered expired.

Default: None, meaning only the authority can close the request.

bounty: An optional fee to reward oracles for priority processing.

Default: 0 lamports.

slots_until_expiration: An optional specification for the max number of slots the request can be processed in.

Default: 2250 slots (roughly 15 minutes at 400 ms/slot).

Minimum: 150 slots (approximately 1 minute at 400 ms/slot).

valid_after_slot: Schedule when the request should be valid for processing.

Default: 0 slots, meaning the request is valid immediately for oracles to process.

signer_seeds: Seeds used for signing.

Returns

ProgramResult: Indicates the result of the instruction execution.

pub fn build_ix( accounts: &FunctionRequestInitAndTriggerAccounts, params: &FunctionRequestInitAndTriggerParams ) -> Result<Instruction, SbError>