Struct load_balancer::balancer::Balancer

pub struct Balancer {
    pub configuration: Config,
    pub servers_count: u64,
    pub next: usize,
    pub client_map: Arc<Mutex<HashMap<String, u64>>>,
    servers: Arc<Mutex<Vec<bool>>>,
}

This structure represents the load balancer itself.

Fields
configuration: Configservers_count: u64next: usizeclient_map: Arc<Mutex<HashMap<String, u64>>>servers: Arc<Mutex<Vec<bool>>>
Implementations

impl Balancer

pub fn new(config: Config) -> Balancer

This function takes a configuration as a parameter and creates the associated load balancer.

async fn check(server: &Server, healthpath: String) -> bool

This function takes a web server as a parameter and returns a boolean. It is responsible for checking whether a specific server is up and healthy or not.

fn client_rate_ok(&mut self, client_address: String) -> bool

This function verifies if a given client (getting it’s IP as a string) has generated too much traffic It returns a boolean indicating if the client can make more requests or not.

pub async fn check_srv(&self) -> bool

This function returns whether the next server in the list is available

pub async fn get_server( &mut self, client_address: String ) -> Result<&Server, &'static str>

This function tries to return a server to the specified client, using the servers list. If this client is blacklisted, the return value will be an error containing a 429 code. If no server is available, the return value will be an error containing a 502 code. Otherwise, a healthy server

async fn check_servers( target_servers: Vec<Server>, healthpath: String ) -> Vec<bool>

This function updates the health status for each server The servers addresses list gets passed as a parameter, as well as the generic healthcheck URL. It returns a vector of booleans indicating whether each server is up and well or not

async fn verify_servers( target_servers: Vec<Server>, servers_status: Arc<Mutex<Vec<bool>>>, interval: u64, healthpath: String )

This function gets started in a new thread and is responsible for starting a healthcheck every n seconds The servers addresses list gets passed as a parameter “target_servers” The servers statuses to update gets passed as “server_status” The number of seconds to wait between each healthcheck is “interval” The healthcheck URL is “healthpath”

async fn clean_clients( client_map: Arc<Mutex<HashMap<String, u64>>>, interval: u64 )

This function is still in development, as the lock for the client_map cant happen as intended for now It should clear the clients map in order to give access to clients that were blocked after a certain amount of time The clients map as well as the time to wait before cleaning it gets passed as parameters

pub fn start_threads(&self)

This function is responsible for starting the two threads “serify_servers” and “clean_clients” For now, the clean_clients function isn’t called as it does not work properly.