chess_rs

Module board

Modules

board_init

Shorthand piece constants to make defining hardcoded board positions in code simpler

Structs

BoardPosition

A zero-indexed position on the chess board

BoardPositionOffset

A type representing an offset in a board position

ChessBoard

A chess board. Does not encode any rules, but is freely changable. Stores the squares such that indexing becomes coordinates, meaning the first array is the first column to the left, and the row index goes upwards. These are natural zero-indexed coordinates.

DeltaChessBoard

A compact type representing a change of a ChessBoard, storing in only 17 bytes what would otherwise take 128

DeltaChessBoardIter

DeltaChessBoardIterMut

Constants

SIDE_LENGTH

The side length of a chess board

Type Aliases

ChessBoardSquare

A square on a chess board