Struct hal_9000::controller::bot::MainController

pub struct MainController {
    pub client: IrcClient,
    pub vote_controller: VoteController,
    pub giveaway_controller: GiveawayController,
    pub filter: Filter,
    pub db_connector: DatabaseConnector,
}

Manages the calls to the other controllers as well as communication between the program and Twitch

client manages IRC communication

vote_controller manages votes in the channel

giveaway_controller manages giveaways in the channel

filter detects swear words in messages

Fields
client: IrcClientvote_controller: VoteControllergiveaway_controller: GiveawayControllerfilter: Filterdb_connector: DatabaseConnector
Methods

impl MainController

pub fn new(lang: String) -> MainController

Instantiates a new MainController

Example

let lang = get_lang();
let client_config: Config = eval_config();
let mut main_controller: MainController = MainController::new(lang, client_config);

pub fn listen(&mut self)

Listens for incoming messages and reacts to them if they match one of the defined patterns