Struct ilockmvp::ilockmvp::ILOCKmvp

pub struct ILOCKmvp {
    psp22: <Data as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    ownable: <Data as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    metadata: <Data as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    reward: <RewardData as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    pool: <TokenPools as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    vest: <VestData as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    app: <AppData as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
}

ILOCKmvp struct contains overall storage data for contract

Fields
psp22: <Data as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type

Openbrush PSP22.

ownable: <Data as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type

Openbrush ownership extension.

metadata: <Data as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type

Openbrush metadata extension.

reward: <RewardData as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type

ILOCK Rewards info.

pool: <TokenPools as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type

ILOCK token pool info.

vest: <VestData as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type

ILOCK vesting info.

app: <AppData as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type

ILOCK connecting application contract info

Implementations

impl ILOCKmvp

pub fn new_token() -> Self

constructor to initialize contract

note: pool contracts must be created prior to construction (for args)

pub fn check_time(&mut self) -> PSP22Result<()>

function to check if enough time has passed to collect next payout

this function ensures Interlock cannot rush the vesting schedule

this function must be called before the next round of token distributions

pub fn remaining_time_until_next_payout( &self ) -> <<ILOCKmvp as ContractEnv>::Env as Environment>::Timestamp

time in seconds until next payout in minutes

pub fn register_stakeholder( &mut self, stakeholder: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, share: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance, pool: u8 ) -> PSP22Result<()>

function that registers a stakeholder’s wallet and vesting info

used to calculate monthly payouts and track net paid

stakeholder data also used for stakeholder to verify their place in vesting schedule

pub fn stakeholder_data( &self, stakeholder: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId ) -> (String, String, String, String)

function that returns a stakeholder’s payout and other data

this will allow stakeholders to verify their stake from explorer if so motivated

returns tuple (paidout, payremaining, payamount, poolnumber)

pub fn distribute_tokens( &mut self, stakeholder: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId ) -> PSP22Result<()>

general function to transfer the token share a stakeholder is currently entitled to

this is called once per stakeholder by Interlock, Interlock paying fees

pools are guaranteed to have enough tokens for all stakeholders

pub fn payout_tokens( &mut self, stakeholder: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, amount: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance, pool: String ) -> PSP22Result<()>

function used to payout tokens to pools with no vesting schedule POOL ARGUMENTS: PARTNERS WHITELIST PUBLIC_SALE

pub fn pool_data(&self, poolnumber: u8) -> (String, String, String, String)

function that returns pool data

this will allow observers to verify vesting parameters for each pool (esp. theirs)

observers may verify pool data from explorer if so motivated

pool numbers range from 0-11

returns (name, tokens, vests, cliff)

pub fn pool_balance( &self, pool: u8 ) -> (String, <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance)

get current balance of whitelist pool

pub fn reward_interlocker( &mut self, reward: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance, interlocker: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId ) -> PSP22Result<<<ILOCKmvp as ContractEnv>::Env as Environment>::Balance>

reward the interlocker for browsing

this is a manual rewarding function, to override the socket formalism

pub fn rewarded_interlocker_total( &self, interlocker: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId ) -> <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance

get amount rewarded to interlocker to date

pub fn rewarded_total( &self ) -> <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance

get total amount rewarded to date

pub fn withdraw_tax( &mut self, wallet: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, amount: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance ) -> PSP22Result<()>

get current balance of whitelist pool

pub fn test_transfer( &mut self, wallet: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, amount: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance ) -> PSP22Result<()>

get current balance of whitelist pool

pub fn tax_available( &self ) -> <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance

display taxpool balance

pub fn months_passed(&self) -> u16

function to get the number of months passed for contract

pub fn cap(&self) -> <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance

function to get the supply cap minted on TGE

pub fn TESTING_increment_month(&mut self) -> bool

function to increment monthspassed for testing

MUST BE DELETED PRIOR TO AUDIT

pub fn update_contract(&mut self, code_hash: [u8; 32]) -> PSP22Result<()>

modifies the code which is used to execute calls to this contract address

this upgrades the token contract logic while using old state

pub fn create_port( &mut self, codehash: <<ILOCKmvp as ContractEnv>::Env as Environment>::Hash, tax: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance, cap: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance, locked: bool, number: u16, owner: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId ) -> PSP22Result<()>

create a new port that rewards contract can register with

eaech port tracks amount rewarded, tax collected, and if it is locked or not

a locked port may only be registered by the interlock network foundation

pub fn create_socket( &mut self, operator: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, portnumber: u16 ) -> OtherResult<()>

rewards/staking contracts register with token contract here

contract must first register with token contract to allow reward transfers

pub fn call_socket( &mut self, address: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, amount: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance, _data: Vec<u8> ) -> OtherResult<()>

check for socket and apply custom logic

pub fn socket( &self, application: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId ) -> Socket

get socket info

pub fn port(&self, portnumber: u16) -> Port

get port info

pub fn emit_event<EE: EmitEvent<Self>>(emitter: EE, event: Event)

function for internal _emit_event implementations

pub fn tax_and_reward( &mut self, address: <<ILOCKmvp as ContractEnv>::Env as Environment>::AccountId, amount: <<ILOCKmvp as ContractEnv>::Env as Environment>::Balance, port: Port, portnumber: u16 ) -> OtherResult<()>

tax and reward socket