Module monkey::lexer

Lexer

The lexer and tokens for Monkey. Has the necessary structs and methods to transform a string to a series of tokens.

Example

use monkey::lexer::{Lexer, Token, Token::{Let, Ident, Assign, Int, Semicolon}};
let program = Lexer::new(String::from("let answer = 42;"));
let tokens = program.into_iter().collect::<Vec<Token>>();
assert_eq!(tokens, vec![Let, Ident("answer".to_owned()), Assign, Int("42".to_owned()), Semicolon]);

Structs

Lexer
The lexer is the workhouse of this module. It's designed to parse a Monkey string into tokens.

Enums

Token
Describes all the tokens for Monkey