Trait acs::logic::unit::Unit

pub trait Unit {
    fn eval(&mut self);

    fn tick(&mut self) { ... }
    fn tock(&mut self) { ... }
}

Common operations for a logic unit.

Calling tick() and tock() determines a cycle, and each cycle is one discrete time unit t.

Required Methods

fn eval(&mut self)

Evaluates the logic expression as we do with arithmetic without advancing the time.

Provided Methods

fn tick(&mut self)

Carries out the tick-phase of a clock cycle: LOW

fn tock(&mut self)

Carries out the tock-phase of a clock cycle: HIGH