Struct acs::mem::fast::ram16k::Ram16k

pub struct Ram16k {
    pub data: Vec<Signal16>,
    /* private fields */
}

Fast RAM which, as opposed to the other RAMs, it does not carries out the simulation with other building blocks, but it emulates it with native code.

Fields
data: Vec<Signal16>
Implementations

impl Ram16k

pub fn sim(
&mut self,
inp: Signal16,
load: Signal,
address: impl Into<Signal14>
) -> Signal16

Carries out the simulation, taking as input:

inp: data input

load: whether to write inp into the RAM

address: which register to select

It returns the value stored at address. Furthermore, when loadisHI, inpis written at locationaddress`, and the loaded value will be emitted from the next time step onward.

pub fn set_inp(&mut self, inp: impl Into<Signal16>)

pub fn set_load(&mut self, load: impl Into<Signal>)

pub fn set_address(&mut self, address: impl Into<Signal14>)

pub fn out(&self) -> Signal16