Module snes::memory

Controls memory IO.

The SNES uses memory maps to access RAM/ROM etc. A lot of the mapping is determined by the cartridge, and there are a number of different mapping structures depending on which cartridge is used. This will eventually serve as an interface to the memory map, and therefore all memory IO should go through here.

Todo

Implement memory mapping

Constants

BITS_IN_BYTE

BITS_IN_WORD

BYTES_IN_LONG

BYTES_IN_WORD

MEMORY_MAX

ONE_BYTE

Functions

get_byte

Memory read

get_word

Memory read

load

Load a binary file into memory

put_byte

Memory write

put_word

Memory write

Type Aliases

Address

BaseAddress

Byte

Word