Module imbriqua_structure_loader::module_output_checker

Tools for cargo checking of a library folder

How to use

Allowing to checking (with cargo) a Rust library file or a Rust executable file, previously genereted.

With purge_folder, this module provide purging folder (as example, before copying content).

With copy_folder, this module provide copying file and folder with availibility check.

With check_result, this module provite cargo command call (test, doc and other).

Minimal usecase

In the case of yout making a package Project_A gererating a Rust library file to test, the TODO list is :

Create a minimal cargo library package (like Imbriqua_Structure_Result, see file tree)

Create code generation script (Project_A objective) writing result in lib.rs file of Project_B

In main of Project_A, use check_result function to test Project_B with cargo

/Project_A/main.rs

mod module_output_checker;
 
fn main() {
 
    fn generate_code ("../Project_B/src/lib.rs") {
        // script part generating file in relative path "../Project_B/src/lib.rs"
    }
     
    module_output_checker::check_result("../Project_B/");
}

Bash cargo equivalent

$ cargo test --manifest-path=".../Project_B/" --all-features --no-run --lib
$ cargo doc --manifest-path=".../Project_B/" --no-deps

File tree

.
├── Project_A/ (executable package, like Imbriqua_Structure_Loader)
│   ├── README.md
│   ├── Cargo.toml
│   ├── Cargo.lock
│   ├── src/
│   │   ├── main.rs
│   │   ├── module_output_checker.rs
│   │   └── ...
│   └── ...
│
├── Project_B/ (library package to check, like Imbriqua_Structure_Result)
│   ├── README.md
│   ├── Cargo.toml
│   ├── Cargo.lock
│   ├── src/
│   │   └── lib.rs
│   └── ...
│
└── ...
 
PS : Of course, you can have a similar folder tree for executable package check

Optimal usecase

In the case of yout making a package Project_A gererating a Rust library file to test, the TODO list is :

Create a minimal cargo library package (like Imbriqua_Structure_Result, see file tree)

Create code generation script (Project_A objective) writing result in a output folder, classified by run time (like versioning folder or archives)

In main of Project_A, use purge_folder function to remove all files in source folder of Project_B (removing old file name)

In main of Project_A, use copy_folder function to copying last result files in source folder of Project_B

In main of Project_A, use check_result function to test Project_B with cargo

/Project_A/main.rs

mod module_output_checker;
 
fn main() {
 
    fn generate_code () {
        // script part generating file and complex folder
        // in relative path "../Output_file/{time_name}/src"
    }
     
    let output_path = format!("../Output_folder/{}", time_name);
 
    // cargo clean, 
    module_output_checker::clean_target_result("../Project_B/");
 
    // copying and checking
    module_output_checker::purge_folder("../Project_B/src/");
    module_output_checker::copy_folder(output_path.as_str(), "../Project_B/src/");
    module_output_checker::check_result("../Project_B/");
 
    // cargo build, for example
    module_output_checker::cargo_custom_command(vec!["build"], "../Project_B/");
}

Bash cargo equivalent

$ cargo test --manifest-path=".../Project_B/" --all-features --no-run --lib
$ cargo doc --manifest-path=".../Project_B/" --no-deps

File tree

.
├── Imbriqua_Structure_Loader/ (executable package)
│   ├── README.md
│   ├── Cargo.toml
│   ├── Cargo.lock
│   ├── src/
│   │   ├── main.rs
│   │   ├── module_output_checker.rs
│   │   └── ...
│   └── ...
│
├── Imbriqua_Structure_Result/ (library package or executable package)
│   ├── README.md
│   ├── Cargo.toml
│   ├── Cargo.lock
│   ├── src/
│   │   ├── lib.rs
│   │   └── lib_folder/ 
│   │       └── ...
│   └── ...
│
├── Output_folder/
│   ├── 2024-04-13_12h36m50/
│   │   ├── lib.rs
│   │   └── lib_folder/
│   │       └── ...
│   ├── 2024-04-12_08h47m01/
│   │   ├── lib.rs
│   │   └── lib_folder/
│   │       └── ...
│   └── ...
│
└── ...

PS : Of course, you can have a similar folder tree for executable package check

Panic and failure

PANIC_OUT01 - The folder don’t exist during purge

Context : module_output_checker.rs/purge_folder()

Info : Can’t find provided folder

Cause : see details in logs file to get :

Value of folder path

PANIC_OUT02 - The folder isn’t readable during purge

Context : module_output_checker.rs/purge_folder()

Info : Can’t read provided folder

Cause : see details in logs file to get :

Value of folder path

error informations of std::path::Path::read_dir()

PANIC_OUT03 - The ‘from’ folder don’t exist (copying)

Context : module_output_checker.rs/copy_folder()

Info : Can’t find provided folder

Cause : see details in logs file to get :

Value of folder path

error informations of std::path::Path::exist()

PANIC_OUT04 - The ‘to’ folder don’t exist (copying)

Context : module_output_checker.rs/copy_folder()

Info : Can’t find provided folder

Cause : see details in logs file to get :

Value of folder path

error informations of std::path::Path::exist()

PANIC_OUT05 - The folder isn’t readable (copying)

Context : module_output_checker.rs/copy_folder()

Info : Can’t read provided folder

Cause : see details in logs file to get :

Value of folder path

error informations of std::path::Path::read_dir()

PANIC_OUT06 - Can’t copying folder

Context : module_output_checker.rs/copy_folder()

Cause : see details in logs file to get :

Value of folder path

error informations of std::fs_extra::copy()

PANIC_OUT07 - Can’t copying file

Context : module_output_checker.rs/copy_folder()

Cause : see details in logs file to get :

Value of folder path

error informations of std::fs::copy()

PANIC_OUT08 - Error in ReadDir iterator

Context : module_output_checker.rs/copy_folder()

Info : Can’t copy provided entry

Cause : see details in logs file to get :

error informations of ReadDir::Iterator

PANIC_OUT09 - Error in OsString::to_str()

Context : module_output_checker.rs/copy_folder()

Info : Can’t copy provided entry

Cause : see details in logs file to get :

Value of entry (debuging syntax)

documentation of ReadDir::Iterator

PANIC_OUT10 - Error in DirEntry::file_type()

Context : module_output_checker.rs/copy_folder()

Info : Can’t copy provided entry

Cause : see details in logs file to get :

error informations of DirEntry::file_type

WARN_OUT01 - Error in ReadDir iterator

Context : module_output_checker.rs/purge_folder()

Info : Can’t remove provided entry

Cause : see details in documentation of ReadDir::Iterator

WARN_OUT02 - Error in pathBuf::to_str()

Context : module_output_checker.rs/purge_folder()

Info : Can’t remove provided entry

Cause : see details in documentation of PathBuf::to_str

WARN_OUT03 - Error in removing entry

Context : module_output_checker.rs/purge_folder()

Info : Can’t remove provided entry

Cause : see details in logs file to get :

Value of folder path

error informations of std::fs_extra::remove_items()

WARN_OUT04 - Cound’t get output

Context : module_output_checker.rs/represent_command_output()

Info : Can’t get output information

Cause : see details in logs file to get :

the command

error informations of std::process::Command::output()

WARN_OUT05 - Couldn’t get STDOUT

Context : module_output_checker.rs/represent_command_output()

Info : Can’t print output information

Cause : see details in logs file to get :

the command

error informations of std::str::from_utf8()

WARN_OUT06 - Couldn’t get STDERR

Context : module_output_checker.rs/represent_command_output()

Info : Can’t print output information

Cause : see details in logs file to get :

the command

error informations of std::str::from_utf8()

Functions

cargo_custom_command

Running cargo custom command

check_result

Checking package of a gived folder, with cargo bash command

clean_target_result

Running cargo clean command

copy_folder

Copy all subelement of a source folder in a target folder

purge_folder

Removing subelement of a target folder

represent_command_output 🔒

Printing command result, used by check_result function