Struct switchboard_solana::oracle_program::accounts::job::JobAccountData

pub struct JobAccountData {
    pub name: [u8; 32],
    pub metadata: [u8; 64],
    pub authority: Pubkey,
    pub expiration: i64,
    pub hash: [u8; 32],
    pub data: Vec<u8>,
    pub reference_count: u32,
    pub total_spent: u64,
    pub created_at: i64,
    pub is_initializing: u8,
}

Fields
name: [u8; 32]

Name of the job to store on-chain.

metadata: [u8; 64]

Metadata of the job to store on-chain.

authority: Pubkey

The account delegated as the authority for making account changes.

expiration: i64

Unix timestamp when the job is considered invalid

hash: [u8; 32]

Hash of the serialized data to prevent tampering.

data: Vec<u8>

Serialized protobuf containing the collection of task to retrieve data off-chain.

reference_count: u32

The number of data feeds referencing the job account..

total_spent: u64

The token amount funded into a feed that contains this job account.

created_at: i64

Unix timestamp when the job was created on-chain.

is_initializing: u8