Struct switchboard_solana::oracle_program::accounts::permission::PermissionAccountData

#[repr(packed(1))]
pub struct PermissionAccountData {
    pub authority: Pubkey,
    pub permissions: u32,
    pub granter: Pubkey,
    pub grantee: Pubkey,
    pub expiration: i64,
    pub bump: u8,
    pub _ebuf: [u8; 255],
}

Fields
authority: Pubkey

The authority that is allowed to set permissions for this account.

permissions: u32

The SwitchboardPermission enumeration assigned by the granter to the grantee.

granter: Pubkey

Public key of account that is granting permissions to use its resources.

grantee: Pubkey

Public key of account that is being assigned permissions to use a granters resources.

expiration: i64

unused currently. may want permission PDA per permission for unique expiration periods, BUT currently only one permission per account makes sense for the infra. Dont over engineer.

bump: u8

The PDA bump to derive the pubkey.

_ebuf: [u8; 255]

Reserved for future info.

Implementations

impl PermissionAccountData

pub fn size() -> usize