Struct switchboard_solana::oracle_program::accounts::queue::OracleQueueAccountData

#[repr(packed(1))]
pub struct OracleQueueAccountData {

    pub name: [u8; 32],
    pub metadata: [u8; 64],
    pub authority: Pubkey,
    pub oracle_timeout: u32,
    pub reward: u64,
    pub min_stake: u64,
    pub slashing_enabled: bool,
    pub variance_tolerance_multiplier: SwitchboardDecimal,
    pub feed_probation_period: u32,
    pub curr_idx: u32,
    pub size: u32,
    pub gc_idx: u32,
    pub consecutive_feed_failure_limit: u64,
    pub consecutive_oracle_failure_limit: u64,
    pub unpermissioned_feeds_enabled: bool,
    pub unpermissioned_vrf_enabled: bool,
    pub curator_reward_cut: SwitchboardDecimal,
    pub lock_lease_funding: bool,
    pub mint: Pubkey,
    pub enable_buffer_relayers: bool,
    pub _ebuf: [u8; 968],
    pub max_size: u32,
    pub data_buffer: Pubkey,

}

Fields
name: [u8; 32]

Name of the queue to store on-chain.

metadata: [u8; 64]

Metadata of the queue to store on-chain.

authority: Pubkey

The account delegated as the authority for making account changes or assigning permissions targeted at the queue.

oracle_timeout: u32

Interval when stale oracles will be removed if they fail to heartbeat.

reward: u64

Rewards to provide oracles and round openers on this queue.

min_stake: u64

The minimum amount of stake oracles must present to remain on the queue.

slashing_enabled: bool

Whether slashing is enabled on this queue.

variance_tolerance_multiplier: SwitchboardDecimal

The tolerated variance amount oracle results can have from the accepted round result before being slashed. slashBound = varianceToleranceMultiplier * stdDeviation Default: 2

feed_probation_period: u32

Number of update rounds new feeds are on probation for. If a feed returns 429s within probation period, auto disable permissions.

curr_idx: u32

Current index of the oracle rotation.

size: u32

Current number of oracles on a queue.

gc_idx: u32

Garbage collection index.

consecutive_feed_failure_limit: u64

Consecutive failure limit for a feed before feed permission is revoked.

consecutive_oracle_failure_limit: u64

Consecutive failure limit for an oracle before oracle permission is revoked.

unpermissioned_feeds_enabled: bool

Enabling this setting means data feeds do not need explicit permission to join the queue and request new values from its oracles.

unpermissioned_vrf_enabled: bool

Enabling this setting means VRF accounts do not need explicit permission to join the queue and request new values from its oracles.

curator_reward_cut: SwitchboardDecimal

TODO: Revenue percentage rewarded to job curators overall.

lock_lease_funding: bool

Prevent new leases from being funded n this queue. Useful to turn down a queue for migrations, since authority is always immutable.

mint: Pubkey

Token mint used for the oracle queue rewards and slashing.

enable_buffer_relayers: bool

Whether oracles are permitted to fulfill buffer relayer update request.

_ebuf: [u8; 968]

Reserved for future info.

max_size: u32

Maximum number of oracles a queue can support.

data_buffer: Pubkey

The public key of the OracleQueueBuffer account holding a collection of Oracle pubkeys that haver successfully heartbeated before the queues oracleTimeout.

Implementations

impl OracleQueueAccountData

pub fn size() -> usize

pub fn convert_buffer(buf: &mut [u8]) -> &mut [Pubkey]

pub fn len(&self) -> u32

pub fn is_empty(&self) -> bool

pub fn get_mint(&self) -> Pubkey

pub fn max_round_rewards(&self, batch_size: u32) -> u64

pub fn new<'info>( account_info: &'info AccountInfo<'info> ) -> Result<Ref<'info, Self>>

Returns the deserialized Switchboard OracleQueue account

Arguments

account_info - A Solana AccountInfo referencing an existing Switchboard OracleQueue

Examples

use switchboard_solana::OracleQueueAccountData;

let oracle_queue = OracleQueueAccountData::new(queue_account_info)?;

pub fn new_from_bytes(data: &[u8]) -> Result<&OracleQueueAccountData>

Returns the deserialized Switchboard OracleQueue account

Arguments

data - A Solana AccountInfo’s data buffer

Examples

use switchboard_solana::OracleQueueAccountData;

let oracle_queue = OracleQueueAccountData::new(oracle_account_info.try_borrow_data()?)?;

pub fn fetch(client: &RpcClient, pubkey: Pubkey) -> Result<Self, SbError>

pub async fn fetch_async( client: &RpcClient, pubkey: Pubkey ) -> Result<Self, SbError>

pub fn fetch_sync<T: SyncClient>( client: &T, pubkey: Pubkey ) -> Result<Self, SbError>

pub async fn fetch_buffer( &self, client: &RpcClient ) -> Result<Vec<Pubkey>, SwitchboardError>

pub async fn fetch_garbage_collection_key( &self, client: &RpcClient ) -> Result<Option<Pubkey>, SwitchboardError>