Struct switchboard_solana::oracle_program::accounts::vrf::VrfAccountData

#[repr(packed(1))]
pub struct VrfAccountData {
    pub status: VrfStatus,
    pub counter: u128,
    pub authority: Pubkey,
    pub oracle_queue: Pubkey,
    pub escrow: Pubkey,
    pub callback: CallbackZC,
    pub batch_size: u32,
    pub builders: [VrfBuilder; 8],
    pub builders_len: u32,
    pub test_mode: bool,
    pub current_round: VrfRound,
    pub _ebuf: [u8; 1024],
}

Fields
status: VrfStatus

The current status of the VRF account.

counter: u128

Incremental counter for tracking VRF rounds.

authority: Pubkey

On-chain account delegated for making account changes.

oracle_queue: Pubkey

The OracleQueueAccountData that is assigned to fulfill VRF update request.

escrow: Pubkey

The token account used to hold funds for VRF update request.

callback: CallbackZC

The callback that is invoked when an update request is successfully verified.

batch_size: u32

The number of oracles assigned to a VRF update request.

builders: [VrfBuilder; 8]

Struct containing the intermediate state between VRF crank actions.

builders_len: u32

The number of builders.

test_mode: boolcurrent_round: VrfRound

Oracle results from the current round of update request that has not been accepted as valid yet

_ebuf: [u8; 1024]

Reserved for future info.

Implementations

impl VrfAccountData

pub fn size() -> usize

pub fn new<'info>( switchboard_vrf: &'info AccountInfo<'_> ) -> Result<Ref<'info, VrfAccountData>>

Returns the deserialized Switchboard VRF account

Arguments

switchboard_vrf - A Solana AccountInfo referencing an existing Switchboard VRF account

Examples

use switchboard_solana::VrfAccountData;

let vrf = VrfAccountData::new(vrf_account_info)?;

pub fn new_from_bytes(data: &[u8]) -> Result<&VrfAccountData>

Returns the deserialized Switchboard VRF account

Arguments

data - A Solana AccountInfo’s data buffer

Examples

use switchboard_solana::VrfAccountData;

let vrf = VrfAccountData::new(vrf_account_info.try_borrow_data()?)?;

pub fn get_current_randomness_round_id(&self) -> u128

Returns the current VRF round ID

pub fn get_result(&self) -> Result<[u8; 32]>

If sufficient oracle responses, returns the latest on-chain result in SwitchboardDecimal format

Examples

use switchboard_solana::VrfAccountData;

pub fn fetch(client: &RpcClient, pubkey: Pubkey) -> Result<Self, SbError>

pub async fn fetch_async( client: &RpcClient, pubkey: Pubkey ) -> Result<Self, SbError>

pub fn fetch_sync<T: SyncClient>( client: &T, pubkey: Pubkey ) -> Result<Self, SbError>