Module switchboard_solana::prelude

Re-exports

pub use crate::SWITCHBOARD_ATTESTATION_PROGRAM_ID;

pub use crate::SWITCHBOARD_ATTESTATION_PROGRAM_ID;

pub use crate::SWITCHBOARD_PROGRAM_ID;

pub use anchor_spl;

pub use rust_decimal;

pub use anchor_client;

pub use anchor_client::anchor_lang;

pub use anchor_client::solana_client;

pub use anchor_client::solana_sdk;

pub use anchor_client::anchor_lang::solana_program;

pub use solana_account_decoder;

pub use futures;

pub use crate::accounts::*;

pub use crate::decimal::*;

pub use crate::error::*;

pub use crate::events::*;

pub use crate::instructions::*;

pub use crate::seeds::*;

pub use crate::types::*;

Modules

NativeMint

The Mint that represents the native token

borsh

error

thiserror

github crates-io docs-rs

Macros

declare_id

Defines the program’s ID. This should be used at the root of all Anchor based programs.

emit

Logs an event that can be subscribed to by clients. Uses the sol_log_data syscall which results in the following log:

err

Returns with the given error. Use this with a custom error type.

error

Generates an Error::AnchorError that includes file and line information.

msg

Print a message to the log.

require

Ensures a condition is true, otherwise returns with the given error. Use this with or without a custom error type.

require_eq

Ensures two NON-PUBKEY values are equal.

require_gt

Ensures the first NON-PUBKEY value is greater than the second NON-PUBKEY value.

require_gte

Ensures the first NON-PUBKEY value is greater than or equal to the second NON-PUBKEY value.

require_keys_eq

Ensures two pubkeys values are equal.

require_keys_neq

Ensures two pubkeys are not equal.

require_neq

Ensures two NON-PUBKEY values are not equal.

source

Creates a Source

Structs

Account

Wrapper around AccountInfo that verifies program ownership and deserializes underlying data into a Rust type.

AccountInfo

Account information

AccountLoader

Type facilitating on demand zero copy deserialization.

AccountMeta

Describes a single account read or written by a program during instruction execution.

AnchorError

AssociatedToken

CacheEntry

Clock

A representation of network time.

Context

Provides non-argument inputs to the program.

CpiContext

Context specifying non-argument inputs for cross-program-invocations.

EpochSchedule

EventHandlerImpl

FunctionResultV0

The schema of the output data that will be sent to the quote verification sidecar.

FunctionResultV1

The schema of the output data that will be sent to the quote verification sidecar.

FunctionResultValidator

Represents a VerifierAccountData oracle and verifies an emitted FunctionResult

FunctionResultValidatorAccounts

The list of accounts used by this verifier to verify the function result

FunctionResultValidatorCache

FunctionResultValidatorInitAccounts

FunctionRunner

An object containing the context to execute a Switchboard Function. Inititlizing this object will load all required variables from the runtime to execute and sign an output transaction to be verified and committed by the switchboard network.

FunctionValidatorVerifyParams

The cleaned up parameters used for a verify instruction

Gramine

Gramine: Gramine is a virtualized runtime used to manage vanilla binaries to execute in an SGX execution environment. This struct allows access to specific overrides that come out-of-the-box with Gramine.

Instruction

A directive for a single invocation of a Solana program.

Instructions

Instructions sysvar, dummy type.

Interface

Type validating that the account is one of a set of given Programs

InterfaceAccount

Wrapper around AccountInfo that verifies program ownership and deserializes underlying data into a Rust type.

Keypair

A vanilla Ed25519 key pair

LegacySolanaFunctionResult

Message

A Solana transaction message (legacy).

Mint

Program

Type validating that the account is the given Program

ProgramData

ProgramErrorWithOrigin

PubSubEventClientBuilder

A builder for creating a new PubSubEventClient

PubSubEventClientWithHandlers

Pubkey

The address of a Solana account.

Rent

Configuration of network rent.

Rewards

ServiceWorkerVerifyMessage

This message is passed by the ServiceWorker to the service to verify the service is permitted to run.

Signer

Type validating that the account signed the transaction. No other ownership or type checks are done. If this is used, one should not try to access the underlying account data.

SlotHashes

SlotHistory

A bitvector indicating which slots are present in the past epoch.

SolanaFunctionEnvironment

The expected environment variables when a solana function is being executed

SolanaFunctionResultV0

Solana Represents the result of a Solana function call.

SolanaFunctionResultV1

Represents the result of a Solana function call.

SolanaFunctionSimulationEnvironment

The expected environment variables when a solana function is being executed

Source

StakeHistory

System

SystemAccount

Type validating that the account is owned by the system program

Sysvar

Type validating that the account is a sysvar and deserializing it.

Token

TokenAccount

Transaction

An atomically-committed sequence of instructions.

TransactionBuilder

Represents a transaction object used for building Solana transactions.

UncheckedAccount

Explicit wrapper for AccountInfo types to emphasize that no checks are performed

Enums

ChainResultInfo

Function result info

Cluster

ComparedValues

Error

ErrorCode

Error codes that can be returned by internal framework code.

ErrorOrigin

FunctionResult

FunctionResultValidatorSigner

LegacyChainResultInfo

ProgramError

Reasons the program may fail

QvnReceipt

Result

Result is a type that represents either success (Ok) or failure (Err).

SbError

Switchboard Functions error suite

SbFunctionError

SolanaFunctionRequestType

SolanaFunctionResult

UpgradeableLoaderState

Upgradeable loader account states

Constants

ERROR_CODE_OFFSET

The starting point for user defined error codes.

Statics

FUNCTION_RESULT_PREFIX

Traits

AccountDeserialize

A data structure that can be deserialized and stored into account storage, i.e. an AccountInfo’s mutable data slice.

AccountSerialize

A data structure that can be serialized and stored into account storage, i.e. an AccountInfo’s mutable data slice.

Accounts

A data structure of validated accounts that can be deserialized from the input to a Solana program. Implementations of this trait should perform any and all requisite constraint checks on accounts to ensure the accounts maintain any invariants required for the program to run securely. In most cases, it’s recommended to use the Accounts derive macro to implement this trait.

AccountsClose

The close procedure to initiate garabage collection of an account, allowing one to retrieve the rent exemption.

AccountsExit

The exit procedure for an account. Any cleanup or persistence to storage should be done here.

AnchorDeserialize

A data-structure that can be de-serialized from binary format by NBOR.

AnchorSerialize

A data-structure that can be serialized into binary format by NBOR.

AsSigner

A trait for types that can act as signers for transactions.

Discriminator

8 byte unique identifier for a type.

EventHandler

EventSenderTrait

Future

A future represents an asynchronous computation obtained by use of async.

Id

Defines the id of a program.

InstructionData

Calculates the data for an instruction invocation, where the data is Sha256(<namespace>:<method_name>)[..8] || BorshSerialize(args). args is a borsh serialized struct of named fields for each argument given to an instruction.

Key

Defines the Pubkey of an account.

Lamports

Lamports related utility methods for accounts.

Owner

Defines an address expected to own an account.

SolanaSysvar

A type that holds sysvar data.

Space

Defines the space of an account for initialization.

ToAccountInfo

Transformation to an AccountInfo struct.

ToAccountInfos

Transformation to AccountInfo structs.

ToAccountMetas

Transformation to AccountMeta structs.

ToKeypair

ZeroCopy

An account data structure capable of zero copy deserialization.

Functions

build_tx

fetch_borsh_account

fetch_borsh_account_async

fetch_borsh_account_sync

fetch_zerocopy_account

Fetches a zero-copy account from the Solana blockchain.

fetch_zerocopy_account_async

Fetches an account asynchronously using the provided client and public key.

fetch_zerocopy_account_sync

Fetches the account data synchronously from the Solana blockchain using the provided client.

generate_signer

Creates a signing keypair generated from randomness sourced from the enclave runtime.

get_async_rpc

get_attestation_program

get_enclave_signer_pubkey

invoke

Invoke a cross-program instruction.

invoke_signed

Invoke a cross-program instruction with program signatures.

ix_to_tx

keypair_from_base_seed

Generates a keypair from a base seed, secret key, optional additional bytes, and an optional program ID.

keypair_from_seed

Constructs a Keypair from caller-provided seed entropy

load_env_pubkey

load_keypair_fs

next_account_info

Convenience function for accessing the next item in an AccountInfo iterator.

parse_optional_pubkey

Parse a string into an optional Pubkey. If the string is empty, return None.

signer_to_pubkey

subscribe

subscribe_ro

subscribe_v2

unix_timestamp

Return the unix timestamp in seconds.

Type Aliases

AnchorClient

AnchorProgram

ProgramResult

QuoteVerifyFn

Attribute Macros

access_control

Executes the given access control method before running the decorated instruction handler. Any method in scope of the attribute can be invoked with any arguments from the associated instruction handler.

account

An attribute for a data structure representing a Solana account.

constant

A marker attribute used to mark const values that should be included in the generated IDL but functionally does nothing.

error_code

Generates Error and type Result<T> = Result<T, Error> types to be used as return types from Anchor instruction handlers. Importantly, the attribute implements From on the ErrorCode to support converting from the user defined error enum into the generated Error.

event

The event attribute allows a struct to be used with emit! so that programs can log significant events in their programs that clients can subscribe to. Currently, this macro is for structs only.

program

The #[program] attribute defines the module containing all instruction handlers defining all entries into a Solana program.

sb_error

switchboard_function

zero_copy

A data structure that can be used as an internal field for a zero copy deserialized account, i.e., a struct marked with #[account(zero_copy)].

Derive Macros

Accounts

Implements an Accounts deserializer on the given struct. Can provide further functionality through the use of attributes.

AnchorDeserialize

AnchorSerialize

InitSpace

Implements a Space trait on the given struct or enum.