Enum switchboard_solana::prelude::ProgramError

pub enum ProgramError {

    Custom(u32),
    InvalidArgument,
    InvalidInstructionData,
    InvalidAccountData,
    AccountDataTooSmall,
    InsufficientFunds,
    IncorrectProgramId,
    MissingRequiredSignature,
    AccountAlreadyInitialized,
    UninitializedAccount,
    NotEnoughAccountKeys,
    AccountBorrowFailed,
    MaxSeedLengthExceeded,
    InvalidSeeds,
    BorshIoError(String),
    AccountNotRentExempt,
    UnsupportedSysvar,
    IllegalOwner,
    MaxAccountsDataAllocationsExceeded,
    InvalidRealloc,
    MaxInstructionTraceLengthExceeded,
    BuiltinProgramsMustConsumeComputeUnits,
    InvalidAccountOwner,
    ArithmeticOverflow,

}

Reasons the program may fail

Variants

Custom(u32)

Allows on-chain programs to implement program-specific error types and see them returned by the Solana runtime. A program-specific error may be any type that is represented as or serialized to a u32 integer.

InvalidArgument

InvalidInstructionData

InvalidAccountData

AccountDataTooSmall

InsufficientFunds

IncorrectProgramId

MissingRequiredSignature

AccountAlreadyInitialized

UninitializedAccount

NotEnoughAccountKeys

AccountBorrowFailed

MaxSeedLengthExceeded

InvalidSeeds

BorshIoError(String)

AccountNotRentExempt

UnsupportedSysvar

IllegalOwner

MaxAccountsDataAllocationsExceeded

InvalidRealloc

MaxInstructionTraceLengthExceeded

BuiltinProgramsMustConsumeComputeUnits

InvalidAccountOwner

ArithmeticOverflow