Struct switchboard_solana::prelude::Signer

pub struct Signer<'info> { /* private fields */ }

Type validating that the account signed the transaction. No other ownership or type checks are done. If this is used, one should not try to access the underlying account data.

Checks:

Signer.info.is_signer == true

Example

#[account]
#[derive(Default)]
pub struct MyData {
    pub data: u64
}

#[derive(Accounts)]
pub struct Example<'info> {
    #[account(init, payer = payer)]
    pub my_acc: Account<'info, MyData>,
    #[account(mut)]
    pub payer: Signer<'info>,
    pub system_program: Program<'info, System>
}

When creating an account with init, the payer needs to sign the transaction.

Implementations

impl<'info> Signer<'info>

pub fn try_from(info: &'info AccountInfo<'info>) -> Result<Signer<'info>, Error>

Deserializes the given info into a Signer.