Struct switchboard_solana::prelude::Sysvar

pub struct Sysvar<'info, T>
where
    T: Sysvar,
{ /* private fields */ }

Type validating that the account is a sysvar and deserializing it.

If possible, sysvars should not be used via accounts but by using the get function on the desired sysvar. This is because using get does not run the risk of Anchor having a bug in its Sysvar type and using get also decreases tx size, making space for other accounts that cannot be requested via syscall.

Example

// OK - via account in the account validation struct
#[derive(Accounts)]
pub struct Example<'info> {
    pub clock: Sysvar<'info, Clock>
}
// BETTER - via syscall in the instruction function
fn better(ctx: Context<Better>) -> Result<()> {
    let clock = Clock::get()?;
}

Implementations

impl<'info, T> Sysvar<'info, T>
where T: Sysvar,

pub fn from_account_info( acc_info: &'info AccountInfo<'info> ) -> Result<Sysvar<'info, T>, Error>