Struct switchboard_solana::prelude::TransactionBuilder

pub struct TransactionBuilder { /* private fields */ }

Represents a transaction object used for building Solana transactions.

The TransactionBuilder struct provides methods for constructing and manipulating Solana transactions. It allows setting the payer, adding instructions, signers, and other transaction parameters. The transaction can be converted to a legacy transaction format or executed directly using the Solana RPC client.

Examples

Creating a new transaction object with a payer:

use solana_sdk::pubkey::Pubkey;
use solana_sdk::instruction::Instruction;
use std::sync::Arc;

let payer = Pubkey::new_unique();
let transaction = TransactionBuilder::new(payer);

Adding an instruction to the transaction:

use solana_sdk::pubkey::Pubkey;
use solana_sdk::instruction::Instruction;
use std::sync::Arc;

let payer = Pubkey::new_unique();
let instruction = Instruction::new_with_bincode(program_id, data);
let transaction = TransactionBuilder::new(payer)
    .add_ix(instruction);

Converting the transaction object to a legacy transaction:

use solana_sdk::pubkey::Pubkey;
use solana_sdk::instruction::Instruction;
use std::sync::Arc;

let payer = Pubkey::new_unique();
let instruction = Instruction::new_with_bincode(program_id, data);
let transaction = TransactionBuilder::new(payer)
    .add_ix(instruction)
    .to_legacy_tx();

Implementations

impl TransactionBuilder

pub fn new(payer: Pubkey) -> Self

pub fn new_with_payer(payer: Arc<dyn AsSigner>) -> Self

pub fn new_with_ixs( payer: Pubkey, ixs: impl IntoIterator<Item = Instruction> ) -> Self

pub fn new_with_payer_and_ixs( payer: Arc<dyn AsSigner>, ixs: impl IntoIterator<Item = Instruction> ) -> Self

pub fn set_compute_units(self, compute_units: u32) -> Self

pub fn set_priority_fees(self, priority_fees: u64) -> Self

pub fn add_ix(self, ix: Instruction) -> Self

pub fn has_signer(&self, signer: Pubkey) -> bool

pub fn has_payer(&self) -> bool

pub fn add_signer(self, signer: Arc<dyn AsSigner>) -> TransactionBuilder

pub fn add_signers(self, signers: Vec<Arc<dyn AsSigner>>) -> TransactionBuilder

pub fn set_recent_blockhash(self, blockhash: Hash) -> Self

pub fn set_min_context_slot(self, min_context_slot: u64) -> Self

pub fn add_address_lookup_account( self, address_lookup_table: AddressLookupTableAccount ) -> Self

pub fn add_address_lookup_accounts( self, address_lookup_tables: &mut Vec<AddressLookupTableAccount> ) -> Self

pub async fn add_address_lookup_table( self, rpc: &RpcClient, address_lookup_table_pubkey: Pubkey ) -> Self

pub async fn add_address_lookup_tables( self, rpc: &RpcClient, address_lookup_table_pubkeys: Vec<Pubkey> ) -> Self

pub fn payer(&self) -> Pubkey

pub fn required_signers(&self) -> Vec<Pubkey>

Return a vec of all of the required signers for the transaction.

pub fn signers(&self) -> Result<Vec<&dyn Signer>, SbError>

Returns the stored signers for the transaction, removing any un-needed signers using the provided ixn’s AccountMeta’s.

pub fn ixs(&self) -> Result<Vec<Instruction>, SbError>

pub fn build_legacy_tx( payer: Pubkey, ixs: Vec<Instruction>, signers: Vec<&dyn Signer>, recent_blockhash: Hash ) -> Result<Transaction, SbError>

pub fn to_legacy_tx(&self) -> Result<Transaction, SbError>

pub fn to_legacy_tx_with_payer<T: AsSigner + Send + Sync>( &self, payer: T ) -> Result<Transaction, SbError>

pub fn to_legacy_tx_with_payer_and_blockhash<T: AsSigner + Send + Sync>( &self, payer: T, recent_blockhash: Option<Hash> ) -> Result<Transaction, SbError>

pub fn to_legacy_tx_with_blockhash( &self, recent_blockhash: Option<Hash> ) -> Result<Transaction, SbError>

pub async fn fetch_address_lookup_account( rpc: &RpcClient, address_lookup_table_pubkey: Pubkey ) -> Result<AddressLookupTableAccount, SbError>

pub async fn fetch_multiple_address_lookup_accounts( rpc: &RpcClient, address_lookup_pubkeys: Vec<Pubkey> ) -> Result<Vec<AddressLookupTableAccount>, SbError>

pub fn build_v0_tx( payer: Pubkey, ixs: Vec<Instruction>, signers: Vec<&dyn Signer>, address_lookup_accounts: Vec<AddressLookupTableAccount>, recent_blockhash: Hash ) -> Result<VersionedTransaction, SbError>

pub fn to_v0_tx(&self) -> Result<VersionedTransaction, SbError>

pub fn to_v0_tx_with_payer<T: AsSigner + Send + Sync>( &self, payer: T ) -> Result<VersionedTransaction, SbError>