Trait switchboard_solana::prelude::InstructionData

pub trait InstructionData: Discriminator + BorshSerialize {
    // Provided method
    fn data(&self) -> Vec<u8>  { ... }
}

Calculates the data for an instruction invocation, where the data is Sha256(<namespace>:<method_name>)[..8] || BorshSerialize(args). args is a borsh serialized struct of named fields for each argument given to an instruction.

Provided Methods

fn data(&self) -> Vec<u8>

Object Safety

This trait is not object safe.