fastlogging

Module root

Statics

PARENT_LOGGER_ADDRESS

ROOT_LOGGER

Functions

add_logger

Add fastlogging logger.

add_writer

add_writer_config

add_writer_configs

add_writers

Add list of writers. writers contains list of writers to add.

critical

Log CRITICAL level message.

debug

Log DEBUG level message.

disable

Disable writer with ID wid.

disable_type

Disable all writers with type typ. See WriterTypeEnum

enable

Enable writer with ID wid.

enable_type

Enable all writers with type typ. See WriterTypeEnum

error

Log ERROR level message.

exception

Log EXCEPTION level message.

fatal

Log FATAL level message.

get_config_string

Get fastlogging configuration as string.

get_parent_client_writer_config

Get IP address of parent process LoggingServer.

get_parent_pid

Get process ID of parent process.

get_parent_pid_client_writer_config

Get process ID of parent process and IP address of parent process LoggingServer.

get_root_server_address_port

get_server_addresses

get_server_addresses_ports

get_server_auth_key

get_server_config

get_server_configs

get_server_ports

get_writer_config

get_writer_configs

info

Log INFO level message.

remove_logger

Remnove fastlogging logger.

remove_writer

remove_writers

Remove list of writer. wids contains list of writer IDs. The return value is a list of removed writers.

root_init

Initialize ROOT logger.

rotate

Rotate a single log file path or all log files with path is None.

save_config

Save fastlogging configuration to file path.

set_debug

Set debug mode.

set_domain

Set logging domain.

set_encryption

set_ext_config

Set extended configuration.

set_level

Set log level for writer with ID wid to level.

set_level2sym

set_root_writer

set_root_writer_config

shutdown

Shutdown fastlogging.

success

Log SUCCESS level message.

sync

Syncronize all writers with types contained in types. Wait for maximum time timeout seconds.

sync_all

Syncronize all writers. Wait for maximum time timeout seconds.

trace

Log TRACE level message.

warning

Log WARNING level message.