Struct keepass::sec_str::SecureString

pub struct SecureString {
    pub string: String,
    // some fields omitted
}

SecureString implements a secure string. This means in particular:

The input string moves to the struct, i.e. it's not just borrowed

The string is encrypted with a random password for obfuscation

mlock() is called on the string to prevent swapping

A method to overwrite the string with zeroes is implemented

The overwrite method is called on drop of the struct automatically

Implements fmt::Show to prevent logging of the secrets, i.e. you can access the plaintext string only via the string value.

Fields

string

Methods

impl SecureString

fn new(string: String) -> SecureString

Create a new SecureString The input string should already lie on the heap, i.e. the type should be String and not &str, otherwise a copy of the plain text string would lie in memory. The string will be automatically encrypted and deleted.

fn delete(&self)

Overwrite the string with zeroes. Call this everytime after unlock() if you don't ned the string anymore.

fn unlock(&mut self)

Unlock the string, i.e. decrypt it and make it available via the string value. Don't forget to call delete() if you don't need the plain text anymore.