Module doc::sorting::insertion

Functions

index_sort

Creates a copy of the list given and sorts it, the given list will remain intact

sort

Insertion sort is one of the simplest sorting algorithms. It is sometimes also known as the brigde hand sorting algorithm. This algorithm is particulary good at sorting partially sorted data, since everything behind index i is sorted. This means that calling sort on a already sorted list, will result in O(N). But the worst case is 1/2N^2=O(N^2).

sort_slice

sorting a slice of a vector lo included hi excluded