Trait hexeng::Recast

pub trait Recast<F: AnyFrame>: Sized {
    fn recast(self, _view_state: &impl ViewTransform) -> F;
}

Implemented by Frame objects, to allow converting to other frames

Required methods

fn recast(self, _view_state: &impl ViewTransform) -> F

Consume the current object and convert it to another frame.

Loading content...