Trait hexeng::ViewTransform

pub trait ViewTransform: Sized {
    type Ref: Deref<Target = Projective2<f64>>;
    fn view_transform(&self) -> Self::Ref;
}

Implemented by objects that can be treated as transformations between world/view coordinates.

Associated Types

type Ref: Deref<Target = Projective2<f64>>

Reference type for the view transform

This can be customised by the implementation to use anything that dereferences to a Projective2<f64> - for example a normal reference or a Box.

Loading content...
Required methods

fn view_transform(&self) -> Self::Ref

Retrieve the view transform associated with this instance.

Loading content...