mcp_daemon::transport
Function handle_ws_connection

pub async fn handle_ws_connection(
    session: Session,
    stream: MessageStream,
    tx: Sender<Message>,
    rx: Receiver<Message>,
) -> Result<()>

Handle a WebSocket connection, managing message flow between client and server

This function sets up bidirectional communication between a WebSocket connection and broadcast channels for message passing.

Arguments

session - The WebSocket session

stream - Stream of incoming WebSocket messages

tx - Channel sender for outgoing messages

rx - Channel receiver for incoming messages

Returns

Result<()> - Ok if the connection was handled successfully, Err otherwise